pyproject.toml
[project]
name = "andreariboni"
version = "0.1.0"
description = "Portfolio homepage"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.14",
    "jinja2>=3.1.6",
    "noise>=1.2.2",
    "numpy>=2.3.1",
    "opencv-python>=4.11.0.86",
    "opensimplex>=0.4.5.1",
    "pillow>=11.2.1",
    "pymbtiles>=0.5.0",
    "scipy>=1.16.0",
    "uvicorn>=0.34.3",
]
